initial (plain) tileset and identified terrain types
This includes a new blank tileset, except for the robot in slot 255.

It creates new variables naming ASCII-art terrain types as placeholders for future tiles.

It changes a few text items relating to Apple Hunt, but is otherwise still Apple Hunt "under the hood." That will be a future commit.