gutted level update mechanism (fixes #9)
there will not be a level progression mechanism in this game, so I removed it / commented it out of the code.

I also had to comment out the updating the bot vars code, because the apple searching code fails when there are no apples. I didn't want to remove that code because we will need that type of functionality later.