fixes #11 and fixes #12
Adds a block of code in key-handling to check if the new "proposed" position is a collision of some kind, which is reversed based on the movement that brought it there. This is to avoid putting all this code in the position modification code above.

I also added code to notice and remark when the robot returns to base.